+++ app/src/main/java/com/makelick/nitrixtest/data/mapper/VideoItemMapper.kt
@@ -5,7 +5,7 @@ import com.makelick.nitrixtest.data.local.model.VideoItem
 import com.makelick.nitrixtest.data.remote.model.VideoCategoryDto
 import com.makelick.nitrixtest.data.remote.model.VideoItemDto
 
-fun VideoItemDto.toEntity(categoryId: Int) = VideoItem(
+fun VideoItemDto.toEntity(categoryId: Long) = VideoItem(
     id = 0,
     categoryId = categoryId,
     description = description,
